pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "td-cli"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
td = "td.cli:cli"
